many fixes for final version

+++ app/controllers/folderController.js
@@ -4,30 +4,45 @@ const aws = require("aws-sdk");
 const folderController = {
   getFolder: async (req, res) => {
     try {
-      const folder = await Folder.findByPk((req.params.folderId), {
-        include: ['pictos']
+      const folder = await Folder.findByPk(req.params.folderId, {
+        include: ["pictos"],
       });
-     res.json(folder);
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+      res.json(folder);
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error.toString());
     }
   },
+
+  getFoldersByOrg: async (req, res) => {
+    try {
+      const folders = await Folder.findAll({
+        where: {
+          org_id: req.params.orgId,
+        },
+      });
+      res.json(folders);
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error.toString());
+    }
+  },
+
   createFolder: async (req, res) => {
     try {
       const newFolder = await Folder.create({
         foldername: req.body.foldername,
-        user_id: req.body.user_id,
+        org_id: req.body.org_id,
         originalname: req.file.originalname,
         mimetype: req.file.mimetype,
         size: req.file.size,
         path: req.file.location,
       });
 
-      res.json(newFolder);
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+      res.json({ newFolder, validation: "Le dossier a bien été créé" });
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error, {error: "Une erreur s'est produite"});
     }
   },
 
@@ -35,60 +50,50 @@ const folderController = {
     const s3 = new aws.S3();
     try {
       const updatedFolder = await Folder.findByPk(req.params.folderId);
-      console.log(updatedFolder);
-      s3.deleteObject(
-        {
-          Bucket: process.env.AWSBucketAv,
-          Key: updatedFolder.originalname,
-        },
-        function (err, _) {
-          if (err) console.log(err, err.stack);
-          else console.log();
-        }
-      );
 
-      const { foldername, user_id } = req.body;
-      const { originalname, mimetype, size, location } = req.file
+      const { foldername, org_id } = req.body;
 
       if (foldername) {
-          foldername
-      }
-      
-      if (user_id) {
-        user_id
+        foldername;
       }
 
-      if (originalname) {
-        originalname
+      if (org_id) {
+        org_id;
       }
 
-      if (mimetype) {
-        mimetype
-      }
+      if (req.file) {
+        s3.deleteObject(
+          {
+            Bucket: process.env.AWSBucketAv,
+            Key: updatedFolder.originalname,
+          },
+          function (err, _) {
+            if (err) console.log(err, err.stack);
+            else console.log();
+          }
+        );
 
-      if (size) {
-        size
-      }
-
-      if (location) {
-        location
+        updatedFolder.set({
+          foldername,
+          org_id,
+          originalname: req.file.originalname,
+          mimetype: req.file.mimetype,
+          size: req.file.size,
+          path: req.file.location,
+        });
       }
 
       updatedFolder.set({
         foldername,
-        user_id,
-        originalname,
-        mimetype,
-        size,
-        path: location
+        org_id,
       });
 
       await updatedFolder.save();
 
-      res.json(updatedFolder);
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+      res.json({ updatedFolder, validation: "Le dossier a été modifié" });
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error, {error: "Une erreur s'est produite"});
     }
   },
 
@@ -107,10 +112,10 @@ const folderController = {
         }
       );
       await deletedFolder.destroy();
-      res.json(deletedFolder);
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+      res.json({ deletedFolder, validation: "Le dossier a bien été supprimé" });
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error, {error: "Une erreur s'est produite"});
     }
   },
 
@@ -124,30 +129,32 @@ const folderController = {
 
       await folder.addPicto(picto);
 
-      res.json({folder, validation: `Le picto a bien été ajouté au favoris ${folder.foldername}`});
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+      res.json({
+        folder,
+        validation: `Le picto a bien été ajouté au favoris ${folder.foldername}`,
+      });
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error, {error: "Une erreur s'est produite"});
     }
   },
 
   removePictoFromFolder: async (req, res) => {
     const { pictoId, folderId } = req.params;
-      console.log(req.body)
+    console.log(req.body);
     try {
-      
       let picto = await Picto.findByPk(pictoId);
 
       let folder = await Folder.findByPk(folderId);
 
       await folder.removePicto(picto);
-     
-      res.json(folder);
-    } catch (err) {
-      console.trace(err);
-      res.status(500).json(err.toString());
+
+      res.json({folder, validation: `Le picto a bien été retiré du favoris ${folder.foldername}`});
+    } catch (error) {
+      console.trace(error);
+      res.status(500).json(error, {error: "Une erreur s'est produite"});
     }
-  }
+  },
 };
 
 module.exports = folderController;
